// frontend/src/openapi/models/SplitMethodEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `ordered` - Ordered
 * * `random` - Random
 * @export
 */
export const SplitMethodEnum = {
    ordered: 'ordered',
    random: 'random'
} as const;
export type SplitMethodEnum = typeof SplitMethodEnum[keyof typeof SplitMethodEnum];


export function instanceOfSplitMethodEnum(value: any): boolean {
    for (const key in SplitMethodEnum) {
        if (Object.prototype.hasOwnProperty.call(SplitMethodEnum, key)) {
            if (SplitMethodEnum[key as keyof typeof SplitMethodEnum] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SplitMethodEnumFromJSON(json: any): SplitMethodEnum {
    return SplitMethodEnumFromJSONTyped(json, false);
}

export function SplitMethodEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): SplitMethodEnum {
    return json as SplitMethodEnum;
}

export function SplitMethodEnumToJSON(value?: SplitMethodEnum | null): any {
    return value as any;
}

export function SplitMethodEnumToJSONTyped(value: any, ignoreDiscriminator: boolean): SplitMethodEnum {
    return value as SplitMethodEnum;
}