// frontend/src/openapi/models/BackendEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `together` - together
 * * `baseten` - baseten
 * * `modal` - modal
 * @export
 */
export const BackendEnum = {
    together: 'together',
    baseten: 'baseten',
    modal: 'modal'
} as const;
export type BackendEnum = typeof BackendEnum[keyof typeof BackendEnum];


export function instanceOfBackendEnum(value: any): boolean {
    for (const key in BackendEnum) {
        if (Object.prototype.hasOwnProperty.call(BackendEnum, key)) {
            if (BackendEnum[key as keyof typeof BackendEnum] === value) {
                return true;
            }
        }
    }
    return false;
}

export function BackendEnumFromJSON(json: any): BackendEnum {
    return BackendEnumFromJSONTyped(json, false);
}

export function BackendEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): BackendEnum {
    return json as BackendEnum;
}

export function BackendEnumToJSON(value?: BackendEnum | null): any {
    return value as any;
}

export function BackendEnumToJSONTyped(value: any, ignoreDiscriminator: boolean): BackendEnum {
    return value as BackendEnum;
}